別ファイルに移動。これで完了
