+++ src/game/window/Window_TitleCommand.ts
@@ -1,0 +1,14 @@
+import { Rectangle } from "src/types/rectangle";
+import { Window_Command } from "./Window_Command";
+
+export declare class Window_TitleCommand extends Window_Command {
+  initialize(rect: Rectangle): void;
+
+  makeCommandList(): void;
+
+  isContinueEnabled(): boolean;
+
+  processOk(): void;
+
+  selectLast(): void;
+}
